+++ changelog.md
@@ -27,6 +27,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityDialogs.DefaultEditAttributeDialogBuilder, bug fixed, retrying a failed update now working again.
 - EntityDialogs.EditAttributeDialogBuilder.Updater removed.
+- EntityDialogs.SelectionDialogBuilder.preferredSize() replaced with dialogSize().
 ### is.codion.swing.framework.ui.test
 - EntityEditPanelTestUnit, createEditPanel() removed, edit panel initializer function parameter added.
 - EntityApplicationPanelTestUnit now initializes all entity panels.